refactor: drain rts-shared no_mangle 404 -> 0
The bulk of docs/specs/no-mangle-drain.md §6 step 3. Every declaration in
rts-shared moves to `#[rtse::abi("<its own symbol>")]`; the linker name is
unchanged everywhere, so no consumer moves. What changes is that 404 symbols
stop being three hand-synced facts (definition / descriptor / Cranelift
signature) and become one derived from the Rust signature — the baked table
grows ~46 KB of derived signatures for rows that previously carried none.

Raw-pointer parameters

  43 of the 404 took a raw `(*const u8, i64)` pair, which has no ABI spelling.
  Each became `(u64, i64)` — the address form the macro's own diagnostic
  sanctions — with the pointer re-introduced as the first statement of the body
  (`let ptr = ptr as *const u8;`), so no function body was touched. Two slots
  before, two slots after: `U64 + I64` and `StrPtr` lower to the same pair of
  i64s, so the machine ABI is byte-identical.

  `&str` would have been the other option and is the WRONG one here: these
  pointers carry bytes out of `Entry::String`, and `&str` would impose UTF-8
  validation the current path does not perform. This is the same reasoning
  abi_sig.rs already records for `__RTS_FN_NS_GC_STRING_NEW`.

  Internal callers of the converted symbols (rts-shared, rts-std, rts-runtime,
  and the crates' own `#[cfg(test)]` code) gained the matching `as u64` at the
  call site.

Verification (spec §5 — every "pre-existing" claim below was produced by
running the failing thing on a stashed, rebuilt clean tree, not inferred)

  cargo check --workspace                    clean
  cargo build --release                      clean
  target/release/rts.exe test                771/774 files, 2837/2849 tests
  cargo test --release -p rts-engine --lib   45 passed
  cargo run -p rts-symbol-baker -- --check   clean
  bash scripts/read_before_commit.sh          no hard violation

The TS failing set is byte-identical to the baseline measured earlier in this
campaign (claude-dom-script-globals 1, claude-object-statics-como-valor crash,
claude-stringify-wrapper-objects 10). No regression.

Measured pre-existing, NOT caused here — both re-run on clean main:
  * `cargo test --release -p rts-codegen-new --lib` fails 7 on main too, with
    the identical test names, including
    `abi_sig::baked_existence::every_named_symbol_is_in_the_baked_table` (its
    own source-extraction sanity check trips: 4 names for 6 arms). abi_sig.rs is
    not touched by this commit.
  * `cargo test -p rts-shared --lib` does not LINK on main either (36 unresolved
    `__rtsadp_*`) — the upward-crate gap §5.3 records for rts-std/rts-node
    covers rts-shared and rts-primitives as well.

+++ crates/rts-runtime/src/adapters/value/regexops.rs
@@ -111,9 +111,9 @@ pub fn rtsadp_re_compile(pat_word: u64, flags_word: u64) -> u64 {
                 explicit.to_string()
             };
             let h = rt_re::__RTS_FN_NS_REGEX_COMPILE(
-                src.as_ptr(),
+                src.as_ptr() as u64,
                 src.len() as i64,
-                flags.as_ptr(),
+                flags.as_ptr() as u64,
                 flags.len() as i64,
             );
             return box_re(h);
@@ -128,9 +128,9 @@ pub fn rtsadp_re_compile(pat_word: u64, flags_word: u64) -> u64 {
     let pat = handle_str(str_handle(pat_word));
     let flags = handle_str(str_handle(flags_word));
     let h = rt_re::__RTS_FN_NS_REGEX_COMPILE(
-        pat.as_ptr(),
+        pat.as_ptr() as u64,
         pat.len() as i64,
-        flags.as_ptr(),
+        flags.as_ptr() as u64,
         flags.len() as i64,
     );
     box_re(h)
@@ -144,7 +144,7 @@ pub fn rtsadp_re_compile(pat_word: u64, flags_word: u64) -> u64 {
 #[rtse::abi]
 pub fn rtsadp_re_test(re_word: u64, subj_word: u64) -> u64 {
     let s = handle_str(str_handle(subj_word));
-    let yes = rt_re::__RTS_FN_NS_REGEX_TEST(unbox_re(re_word), s.as_ptr(), s.len() as i64) != 0;
+    let yes = rt_re::__RTS_FN_NS_REGEX_TEST(unbox_re(re_word), s.as_ptr() as u64, s.len() as i64) != 0;
     PolyValue::bool(yes).raw()
 }
 
@@ -213,7 +213,7 @@ pub fn rtsadp_re_str_match(subj_word: u64, re_word: u64) -> u64 {
     }
     let s = handle_str(str_handle(subj_word));
     let raw_vec =
-        rt_re::__RTS_FN_NS_REGEX_MATCH_GROUPS(unbox_re(re_word), s.as_ptr(), s.len() as i64);
+        rt_re::__RTS_FN_NS_REGEX_MATCH_GROUPS(unbox_re(re_word), s.as_ptr() as u64, s.len() as i64);
     if raw_vec == 0 {
         return PolyValue::null().raw();
     }
@@ -508,6 +508,6 @@ pub fn rtsadp_re_str_split(subj_word: u64, re_word: u64, limit_word: u64) -> u64
 #[rtse::abi]
 pub fn rtsadp_re_str_search(subj_word: u64, re_word: u64) -> u64 {
     let s = handle_str(str_handle(subj_word));
-    let idx = rt_re::__RTS_FN_NS_REGEX_FIND_AT(unbox_re(re_word), s.as_ptr(), s.len() as i64);
+    let idx = rt_re::__RTS_FN_NS_REGEX_FIND_AT(unbox_re(re_word), s.as_ptr() as u64, s.len() as i64);
     genops::number_result(idx as f64).raw()
 }
